feat(AppShell): add WithMobileNav story, revert MobileNav stories
Move the app shell integration story from MobileNav.stories to
AppShell.stories where it belongs — this is a layout pattern, not
a MobileNav demo. Reverts MobileNav.stories to its original state.
